pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "plexichat"
version = "1.0.0"
description = "A comprehensive chat platform with advanced features"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "PlexiChat Team", email = "team@plexichat.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
requires-python = ">=3.8"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-multipart>=0.0.6",
    "bcrypt>=4.0.1",
    "psutil>=5.9.0",
    "websockets>=12.0",
    "aiofiles>=23.2.1",
    "python-dotenv>=1.0.0",
    "cryptography>=41.0.7",
    "sqlite-utils>=3.35.2",
    "sqlalchemy[asyncio]>=2.0.23",
    "aiosqlite>=0.19.0",
    "asyncio-mqtt>=0.13.0",
    "httpx>=0.25.2",
    "rich>=13.7.0",
    "typer>=0.9.0",
    "colorama>=0.4.6",
    "click>=8.1.7",
    "Pillow>=10.1.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.2",
    "lxml>=4.9.3",
    "markdown>=3.5.1",
    "PyYAML>=6.0.1",
    "toml>=0.10.2",
    "python-dateutil>=2.8.2",
    "packaging>=23.2",
    "setuptools>=69.0.0",
    "wheel>=0.42.0",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "pyotp>=2.9.0",
    "qrcode[pil]>=7.4.2",
    "email-validator>=2.1.0",
    "tabulate>=0.9.0",
    "motor>=3.3.2",
    "aioredis>=2.0.1",
    "aiohttp>=3.9.1",
    "aiokafka>=0.10.0",
    "aio-pika>=9.3.1",
    "jsonschema>=4.20.0",
    "werkzeug>=3.0.1",
    "flask>=3.0.0",
    "pefile>=2023.2.7",
    "phonenumbers>=8.13.27",
    "psycopg2-binary>=2.9.9",
    "aiomysql>=0.2.0",
    "alembic>=1.13.1",
    "transformers>=4.36.0",
    "torch>=2.1.1",
    "scikit-learn>=1.3.2",
    "numpy>=1.24.4",
    "pandas>=2.1.4",
    "matplotlib>=3.8.2",
    "seaborn>=0.13.0",
    "plotly>=5.17.0",
    "opencv-python>=4.8.1.78",
    "imageio>=2.33.0",
    "selenium>=4.16.0",
    "scrapy>=2.11.0",
    "redis>=5.0.1",
    "celery>=5.3.4",
    "paramiko>=3.4.0",
    "fabric>=3.2.2",
    "jinja2>=3.1.2",
    "openpyxl>=3.1.2",
    "xlsxwriter>=3.1.9",
    "schedule>=1.2.0",
    "watchdog>=3.0.0",
    "python-socketio>=5.10.0",
    "eventlet>=0.33.3",
    "customtkinter>=5.2.0",
    "tkinter-tooltip>=2.1.0",
    "openai>=1.0.0",
    "PyQt6>=6.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]

[project.urls]
Homepage = "https://github.com/plexichat/plexichat"
Documentation = "https://docs.plexichat.com"
Repository = "https://github.com/plexichat/plexichat"
"Bug Tracker" = "https://github.com/plexichat/plexichat/issues"

[project.scripts]
plexichat = "plexichat.main:main"
plexichat-cli = "plexichat.interfaces.cli.main_cli:main"
plexichat-gui = "plexichat.interfaces.gui:main"
plexichat-web = "plexichat.interfaces.web.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"*" = ["*.html", "*.css", "*.js", "*.json", "*.yaml", "*.yml", "*.toml", "*.md", "*.txt"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
